+++ tests/FakeXrmEasy.Plugins.Tests/IXrmFakedContextPipelineExtensions/GetStepsForStageWithRetrieveMultipleTests.cs
@@ -158,8 +158,9 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
             Assert.Equal(pluginType.Assembly.GetName().Name, pluginStep.AssemblyName);
             Assert.Equal(pluginType.FullName, pluginStep.PluginType);
 
-            Assert.NotEmpty(pluginStep.FilteringAttributes);
-            Assert.Equal("name", pluginStep.FilteringAttributes[0]);
+            var attributes = pluginStep.FilteringAttributes.ToList();
+            Assert.NotEmpty(attributes);
+            Assert.Equal("name", attributes[0]);
         }
 
         [Theory]
@@ -187,9 +188,10 @@ namespace FakeXrmEasy.Plugins.Tests.IXrmFakedContextPipelineExtensions
             Assert.Equal(pluginType.Assembly.GetName().Name, pluginStep.AssemblyName);
             Assert.Equal(pluginType.FullName, pluginStep.PluginType);
 
-            Assert.NotEmpty(pluginStep.FilteringAttributes);
-            Assert.Equal("name", pluginStep.FilteringAttributes[0]);
-            Assert.Equal("accountcategorycode", pluginStep.FilteringAttributes[1]);
+            var attributes = pluginStep.FilteringAttributes.ToList();
+            Assert.NotEmpty(attributes);
+            Assert.Equal("name", attributes[0]);
+            Assert.Equal("accountcategorycode", attributes[1]);
         }
 
         [Theory]