+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Added 
 
+ - Plugin Step registration improvements: new general purpose method to register plugin steps: #53
  - Added possibility to register plugin steps with logical name, and therefore support for late bound registration #38
- - 
+ 
 ### Changed
 
 - Fix Sonar Quality Gate settings: DynamicsValue/fake-xrm-easy#28